Rename controller files
The files containing the controllers were renamed to index.controller.ts
